+++ lightfm/__init__.py
@@ -1,1 +1,5 @@
 from .lightfm import LightFM
+
+__version__ = '1.9'
+
+__all__ = ['LightFM', 'evaluation']