+++ lightfm/datasets/__init__.py
@@ -1,0 +1,1 @@
+from lightfm.datasets.movielens import fetch_movielens